enum classを別ファイルに移動
